- Added more colors
- Fixed various bugs
- Added line clipping
- Fixed various warnings
- Added KeyDown event
- Added cube test

+++ Src/3DGame/Engine/MathEx.h
@@ -18,7 +18,7 @@
 typedef unsigned char octet;
 typedef unsigned int uint;
 
-const float deg2rad = M_PI / 180.0f;
+const float deg2rad = float(M_PI / 180.0f);
 
 float barycentric(float v1, float v2, float v3, float b2, float b3);
 float catmullRom(float v1, float v2, float v3, float v4, float a);
